refactor(agentadapters): split acp_session.go into per-type files
acp_session.go held 1651 LOC and 73 functions across 4 types
(SessionManager, acpSession, acpChatClient, acpTurn). One type
per file mirrors the surrounding agentadapters layout.

  acp_session.go:             1651 → 85 LOC (constants + Runner interface)
  session_manager.go:         new, 324 LOC (SessionManager + methods)
  acp_session_lifecycle.go:   new, 352 LOC (acpSession + lifecycle methods)
  acp_chat_client.go:         new, 185 LOC (acpChatClient + ACP RPC handlers)
  acp_turn.go:                new, 805 LOC (acpTurn + buffering / dedupe helpers)

What stays in acp_session.go

  The shared package-level surface: `package agentadapters`,
  imports, `DriverKindACP`, `ErrSessionNotActive`, and the public
  `Runner` interface. Type-specific constants moved with their
  consumers (thoughtFallbackBlockID / thoughtMaxBytesPerBlock /
  thoughtTruncationSuffix live with acp_turn.go's acpTurn;
  acpShutdownCancelTimeout / acpShutdownCloseTimeout live with
  session_manager.go's SessionManager).

What moves to session_manager.go

  SessionManager struct + sessionStart helper + 15 methods:
  NewSessionManager, SetLogger, SetApprovalCoordinator,
  SetAdapterMetrics, SetCredentialEnvProvider, SetShutdownCancelHook,
  Coordinator, PrepareSession, Run, session, CloseSession,
  Shutdown, SetSessionConfigOption.

What moves to acp_session_lifecycle.go

  acpSession struct + lifecycle methods: startACPSession,
  RunTurn, SetConfigOption, setConfigOptions, configOptionsSnapshot,
  cloneConfigOptions, Close, setActiveTurn, clearActiveTurn,
  cancelActiveTurn, captureACPTurnResult.

What moves to acp_chat_client.go

  acpChatClient struct + the ACP reverse-RPC handlers it owns:
  setTurn, clearTurn, currentTurn, SessionUpdate, RequestPermission,
  ReadTextFile, WriteTextFile, CreateTerminal, KillTerminal,
  TerminalOutput, ReleaseTerminal, WaitForTerminalExit,
  workspacePath, terminalRPCUnsupported.

What moves to acp_turn.go

  acpTurn struct + the per-turn accumulator and helpers:
  newACPTurn, setActivityCallback, recordUpdate, appendAgentMessageChunk,
  appendAgentThoughtChunk, appendBoundedThoughtText, recordToolCall,
  recordToolCallUpdate, rememberToolKind, lookupToolKind,
  emitFileChangeActivities, summarizeFileChangeLines,
  isFileMutatingToolKind, markToolSeen, isLikelyProgressNarration,
  recordPlan, emitActivity, acpToolStatus, toolCallDetail,
  summarizeToolRawInput, flattenRawInput, firstRawInputValue,
  firstRawInputString, summarizeToolLocations, summarizeToolContent,
  trimToolSummary, pluralize, firstNonEmpty, recordUsage, appendRaw,
  snapshot, truncated, contentBlockText, stderrSuffix,
  terminateProcess.

No exported symbol moves to a different package; all 4 new files
stay `package agentadapters`; the test file
(acp_session_test.go, 51.7 KB) stays as one to limit risk.
Imports per file are minimal, pruned via goimports.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./internal/agentadapters/... -race -count=1  — 226 / 226 pass
  go test ./... -race -count=1  — 1779 / 1779 pass
